+++ src/Okanshi/Monitor.fs
@@ -10,37 +10,25 @@ type MonitorKey = (string * Type * string)
 /// Type representing a monitor factory
 type MonitorFactory = unit -> IMonitor
 
-type OkanshiMonitorMessage =
-    private
-    | GetMonitor of MonitorKey * MonitorFactory * AsyncReplyChannel<IMonitor>
-
 /// Static use of monitors
 [<AbstractClass; Sealed>]
 type OkanshiMonitor private() =
-    static let cancellationTokenSource = new CancellationTokenSource()
-    static let cancellationToken = cancellationTokenSource.Token
-
-    static let monitorAgentLoop (inbox : MailboxProcessor<OkanshiMonitorMessage>) =
-        let registeredMonitors = new Dictionary<MonitorKey, IMonitor>()
-        let monitorRegistry = DefaultMonitorRegistry.Instance
-        let rec loop () =
-            async {
-                let! msg = inbox.Receive()
-                match msg with
-                | GetMonitor (key, factory, reply) ->
-                    match registeredMonitors.TryGetValue(key) with
-                    | success, monitor when success -> reply.Reply(monitor)
-                    | success, _ when not success ->
-                        let monitor = factory()
-                        registeredMonitors.Add(key, monitor)
-                        monitorRegistry.Register(monitor)
-                        reply.Reply(monitor)
-                    | _ -> ()
-
-                return! loop()
-            }
-        loop()
-    static let monitorAgent = MailboxProcessor.Start(monitorAgentLoop, cancellationToken)
+    static let registeredMonitors = new Dictionary<MonitorKey, IMonitor>()
+    static let registeredMonitorsLock = new obj()
+    static let monitorRegistry = DefaultMonitorRegistry.Instance
+        
+    static let getOrAddMonitor monitorKey (factory : MonitorFactory) =
+        // It is a conscious to use a lock instead of a ConcurrentDictionary. This is due to the fact that the
+        // factory method in ConcurrentDictionary.GetOrAdd, aren't thread-safe, it may be called multiple times
+        // which would cause the monitor to be registered multiple times.
+        lock registeredMonitorsLock (fun () ->
+            match registeredMonitors.TryGetValue(monitorKey) with
+            | true, result -> result
+            | false, _ ->
+                let monitor = factory()
+                registeredMonitors.Add(monitorKey, monitor)
+                monitorRegistry.Register(monitor)
+                monitor)
 
     static let tagDictionary = new ConcurrentDictionary<Tag, byte>()
     static let defaultStep = TimeSpan.FromMinutes(float 1)
@@ -66,7 +54,7 @@ type OkanshiMonitor private() =
     static member BasicCounter(name : string, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<BasicCounter>, tags)
         let factory = fun () -> (new BasicCounter(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags)) :> IMonitor)
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> BasicCounter
+        getOrAddMonitor monitorKey factory :?> BasicCounter
 
     /// Get or add a StepCounter, with a step size of 1 minute
     static member StepCounter(name) = OkanshiMonitor.StepCounter(name, defaultStep, [||])
@@ -78,7 +66,7 @@ type OkanshiMonitor private() =
     static member StepCounter(name : string, step, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<StepCounter>, tags)
         let factory = fun () -> (new StepCounter(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags), step) :> IMonitor)
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> StepCounter
+        getOrAddMonitor monitorKey factory :?> StepCounter
 
     /// Get or add a PeakRateCounter, with a step size of 1 minute
     static member PeakRateCounter(name) = OkanshiMonitor.PeakRateCounter(name, defaultStep, [||])
@@ -90,7 +78,7 @@ type OkanshiMonitor private() =
     static member PeakRateCounter(name : string, step, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<PeakRateCounter>, tags)
         let factory = fun () -> (new PeakRateCounter(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags), step) :> IMonitor)
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> PeakRateCounter
+        getOrAddMonitor monitorKey factory :?> PeakRateCounter
 
     /// Get or add a DoubleCounter, with a step size of 1 minute
     static member DoubleCounter(name) = OkanshiMonitor.DoubleCounter(name, defaultStep, [||])
@@ -102,7 +90,7 @@ type OkanshiMonitor private() =
     static member DoubleCounter(name : string, step, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<DoubleCounter>, tags)
         let factory = fun () -> (new DoubleCounter(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags), step) :> IMonitor)
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> DoubleCounter
+        getOrAddMonitor monitorKey factory :?> DoubleCounter
 
     /// Get or add a BasicGauge
     static member BasicGauge(name : string, getValue : Func<'T>) = OkanshiMonitor.BasicGauge(name, getValue, [||])
@@ -111,7 +99,7 @@ type OkanshiMonitor private() =
     static member BasicGauge<'T>(name : string, getValue : Func<'T>, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<BasicGauge<'T>>, tags)
         let factory = fun () -> (new BasicGauge<'T>(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags), getValue)) :> IMonitor
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> BasicGauge<'T>
+        getOrAddMonitor monitorKey factory :?> BasicGauge<'T>
 
     /// Get or add a MaxGauge 
     static member MaxGauge(name : string) = OkanshiMonitor.MaxGauge(name, [||])
@@ -120,7 +108,7 @@ type OkanshiMonitor private() =
     static member MaxGauge(name : string, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<MaxGauge>, tags)
         let factory = fun () -> (new MaxGauge(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags))) :> IMonitor
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> MaxGauge
+        getOrAddMonitor monitorKey factory :?> MaxGauge
 
     /// Get or add a MinGauge
     static member MinGauge(name : string) = OkanshiMonitor.MinGauge(name, [||])
@@ -129,7 +117,7 @@ type OkanshiMonitor private() =
     static member MinGauge(name : string, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<MinGauge>, tags)
         let factory = fun () -> (new MaxGauge(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags))) :> IMonitor
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> MinGauge
+        getOrAddMonitor monitorKey factory :?> MinGauge
 
     /// Get or add a LongGauge
     static member LongGauge(name : string) = OkanshiMonitor.LongGauge(name, [||])
@@ -138,7 +126,7 @@ type OkanshiMonitor private() =
     static member LongGauge(name : string, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<LongGauge>, tags)
         let factory = fun () -> (new LongGauge(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags))) :> IMonitor
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> LongGauge
+        getOrAddMonitor monitorKey factory :?> LongGauge
 
     /// Get or add a DoubleGauge
     static member DoubleGauge(name : string) = OkanshiMonitor.DoubleGauge(name, [||])
@@ -147,7 +135,7 @@ type OkanshiMonitor private() =
     static member DoubleGauge(name : string, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<DoubleGauge>, tags)
         let factory = fun () -> (new DoubleGauge(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags))) :> IMonitor
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> DoubleGauge
+        getOrAddMonitor monitorKey factory :?> DoubleGauge
 
     /// Get or add a DecimalGauge
     static member DecimalGauge(name : string) = OkanshiMonitor.DecimalGauge(name, [||])
@@ -156,7 +144,7 @@ type OkanshiMonitor private() =
     static member DecimalGauge(name : string, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<DecimalGauge>, tags)
         let factory = fun () -> (new DecimalGauge(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags))) :> IMonitor
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> DecimalGauge
+        getOrAddMonitor monitorKey factory :?> DecimalGauge
 
     /// Get or add a BasicTimer, with a step size of 1 minute
     static member BasicTimer(name) = OkanshiMonitor.BasicTimer(name, defaultStep, [||])
@@ -168,7 +156,7 @@ type OkanshiMonitor private() =
     static member BasicTimer(name : string, step, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<BasicTimer>, tags)
         let factory = fun () -> (new BasicTimer(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags), step)) :> IMonitor
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> BasicTimer
+        getOrAddMonitor monitorKey factory :?> BasicTimer
 
     /// Get or add a DurationTimer
     static member LongTaskTimer(name) = OkanshiMonitor.LongTaskTimer(name, [||])
@@ -177,7 +165,7 @@ type OkanshiMonitor private() =
     static member LongTaskTimer(name : string, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<LongTaskTimer>, tags)
         let factory = fun () -> (new LongTaskTimer(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags))) :> IMonitor
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> LongTaskTimer
+        getOrAddMonitor monitorKey factory :?> LongTaskTimer
 
     /// Get or add a HealthCheck
     static member HealthCheck(check, name) = OkanshiMonitor.HealthCheck(check, name)
@@ -186,7 +174,7 @@ type OkanshiMonitor private() =
     static member HealthCheck(check, name : string, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<HealthCheck>, tags)
         let factory = fun () -> (new HealthCheck(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags), check)) :> IMonitor
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> HealthCheck
+        getOrAddMonitor monitorKey factory :?> HealthCheck
 
     /// Get or add a performance counter monitor
     static member PerformanceCounter(check, name) = OkanshiMonitor.PerformanceCounter(check, name)
@@ -195,4 +183,4 @@ type OkanshiMonitor private() =
     static member PerformanceCounter(counterConfig : PerformanceCounterConfig, name, tags : Tag array) =
         let monitorKey = OkanshiMonitor.GetMonitorKey(name, typeof<PerformanceCounterMonitor>, tags)
         let factory = fun () -> (new PerformanceCounterMonitor(MonitorConfig.Build(name).WithTags(OkanshiMonitor.DefaultTags).WithTags(tags), counterConfig)) :> IMonitor
-        monitorAgent.PostAndReply(fun reply -> GetMonitor(monitorKey, factory, reply)) :?> PerformanceCounterMonitor
+        getOrAddMonitor monitorKey factory :?> PerformanceCounterMonitor