+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 * Fix dependencies on NuGet package
 * Add new NuGet package with support for OWIN.
 * Added support for registering a custom timing by providing a number describing the elapsed number of milliseconds.
+* Improved performance when using OkanshiMonitor to get monitors
 
 ### 4.0.0-beta